Recent updates 12/13/2018

+++ src/main/java/gwt/material/design/amcharts/client/properties/RadarColumnProperties.java
@@ -27,5 +27,5 @@ import jsinterop.annotations.JsType;
 public class RadarColumnProperties extends ColumnProperties {
 
     @JsProperty
-    public int cornerRadius;
+    public double cornerRadius;
 }
